dynamic groups action fixes
