+++ app/packages/components/src/components/Loading/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./Loading";
+export { default as LoadingDots } from "./LoadingDots";
