+++ app/packages/components/src/components/index.ts
@@ -1,14 +1,18 @@
 export { default as Bar } from "./Bar";
 export { default as Button } from "./Button";
+export { default as CodeBlock } from "./CodeBlock";
+export { default as CopyButton } from "./CopyButton";
 export { default as ErrorBoundary } from "./ErrorBoundary";
 export { default as ExternalLink, useExternalLink } from "./ExternalLink";
+export { default as FilterAndSelectionIndicator } from "./FilterAndSelectionIndicator";
 export { default as Header } from "./Header";
 export { default as HelpPanel } from "./HelpPanel";
 export { default as IconButton } from "./IconButton";
 export * from "./Icons";
 export { default as JSONPanel } from "./JSONPanel";
 export { default as Link } from "./Link";
-export { default as Loading } from "./Loading";
+export { default as Loading, LoadingDots } from "./Loading";
+export { default as PillButton } from "./PillButton";
 export { default as Popout } from "./Popout";
 export { default as PopoutSectionTitle } from "./PopoutSectionTitle";
 export { default as RouteRenderer } from "./RouteRenderer";
@@ -17,7 +21,3 @@ export { default as Selector } from "./Selector";
 export { default as TabOption } from "./TabOption";
 export { default as ThemeProvider, useTheme } from "./ThemeProvider";
 export { default as Tooltip } from "./Tooltip";
-export { default as PillButton } from "./PillButton";
-export { default as FilterAndSelectionIndicator } from "./FilterAndSelectionIndicator";
-export { default as CopyButton } from "./CopyButton";
-export { default as CodeBlock } from "./CodeBlock";
