+++ app/packages/relay/src/queries/index.ts
@@ -1,25 +1,26 @@
+export * from "./__generated__/aggregateQuery.graphql";
 export * from "./__generated__/aggregationsQuery.graphql";
 export * from "./__generated__/configQuery.graphql";
 export * from "./__generated__/countValuesQuery.graphql";
 export * from "./__generated__/histogramValuesQuery.graphql";
 export * from "./__generated__/mainSampleQuery.graphql";
-export * from "./__generated__/paginateGroup_query.graphql";
+export * from "./__generated__/paginateDynamicGroupSamplesQuery.graphql";
 export * from "./__generated__/paginateGroupPageQuery.graphql";
 export * from "./__generated__/paginateGroupQuery.graphql";
-export * from "./__generated__/paginateDynamicGroupSamplesQuery.graphql";
+export * from "./__generated__/paginateGroup_query.graphql";
 export * from "./__generated__/pcdSampleQuery.graphql";
 export * from "./__generated__/viewSchemaFragment.graphql";
 export * from "./__generated__/viewSchemaFragmentQuery.graphql";
-
+export { default as aggregate } from "./aggregate";
 export { default as aggregation } from "./aggregations";
 export { default as config } from "./config";
 export { default as countValues } from "./countValues";
 export { default as histogramValues } from "./histogramValues";
 export { default as mainSample } from "./mainSample";
+export { default as paginateDynamicGroupSamples } from "./paginateDynamicGroupSamples";
 export {
   default as paginateGroup,
   paginateGroupPaginationFragment,
 } from "./paginateGroup";
-export { default as paginateDynamicGroupSamples } from "./paginateDynamicGroupSamples";
 export { default as pcdSample } from "./pcdSample";
 export { default as viewSchema } from "./viewSchema";
