+++ fiftyone/server/routes/__init__.py
@@ -6,7 +6,6 @@ FiftyOne Server routes
 |
 """
 from .aggregate import Aggregate
-from .dynamic_groups import DynamicGroupsRoutes
 from .embeddings import EmbeddingsRoutes
 from .event import Event
 from .events import Events
@@ -26,8 +25,7 @@ from fiftyone.operators.server import OperatorRoutes
 
 # Starlette routes should not be created here. Please leave as tuple definitions
 routes = (
-    DynamicGroupsRoutes
-    + EmbeddingsRoutes
+    EmbeddingsRoutes
     + OperatorRoutes
     + [
         ("/aggregate", Aggregate),
